Options and Music
Added functionality to the Options Tab and added music to the game. The Music is a non copyright file of a medieval theme that plays constant loop even in the game. In the options menu theres the option to lower the volume in game for the music. This is done using an audio mixer group linked to the option's sound slider. I also added a glowing effect to the shard on the main menu screen.