Made the Pause menu, esc key to enter pause
This pause menu contains pause game, options, back to main menu, and save game. The save game portion is still bugged it shows the saved time slot in the save slot when the save was made however when trying to load the save the player will end up at the start of level one every time.