More save slots
When the user clicks on new game they'r prompted to choose a save slot from the start. If there's already a save on the chosen slot a warning with pop up asking if the user would like to delete that current save. Theres some things I want to fix being that If you want to save in game you'll get the overwrite pop up, I don't think you should when you're currently playing on the selected slot. Also forgot to add a warning for leaving the game before saving.